pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "doorbell"
version = "0.1.0"
description = "Raspberry Pi doorbell system with SIP calling and video streaming"
readme = "README.md"

# Replace classifier with SPDX license
license = { text = "MIT" }

authors = [
    {name = "Bernhard Rode", email = "mail@bernhardrode.de"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "rpi-lgpio>=0.6",
    "uvicorn>=0.35.0",
    "aiortc>=1.9.0",
    "starlette>=0.40.0",
    "pyaudio>=0.2.11",
    "pysipio>=1.8.0",
    "scipy>=1.16.1",
]

[project.urls]
Homepage = "https://github.com/bernhardrode/doorbell"
Repository = "https://github.com/bernhardrode/doorbell"
Issues = "https://github.com/bernhardrode/doorbell/issues"

[project.scripts]
doorbell = "doorbell.main:main"
doorbell-install-service = "install_service:install_systemd_service"
doorbell-uninstall-service = "uninstall_service:uninstall_systemd_service"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["doorbell"]
artifacts = ["doorbell.service", "install_service.py", "uninstall_service.py"]
